requirements.txt
agents
openai
python-dotenv
jinja2
pydantic 